gcf-scan-write-storage-snow/requirements.txt
functions-framework
google-cloud-storage
avro-python3
python-snappy